better cache fiddling, minor changes

+++ mongo.cpp
@@ -36,7 +36,7 @@ void lock_internal::lock(const std::string& debug_info, size_t who, mongoc_clien
     //size_t cur_retry = 0;
 
     ///200 ms
-    constexpr size_t max_microseconds_elapsed = 1000 * 200;
+    constexpr int64_t max_microseconds_elapsed = 1000 * 200;
     bool sleeptime = false;
 
     sf::Clock clk;
